update nearly done
delete on cascade still problem

+++ application/views/klien.php
@@ -6,8 +6,8 @@ if (!empty($this->session->userdata('id_pengguna'))) {
 <html lang="en">
 <head>
     <meta charset="utf-8" />
-    <link rel="apple-touch-icon" sizes="76x76" href="<?php echo base_url(); ?>/assets/img/apple-icon.png">
-    <link rel="icon" type="image/png" sizes="96x96" href="<?php echo base_url(); ?>/assets/img/favicon.png">
+    <link rel="apple-touch-icon" sizes="76x76" href="<?php echo base_url(); ?>assets/img/apple-icon.png">
+    <link rel="icon" type="image/png" sizes="96x96" href="<?php echo base_url(); ?>assets/img/favicon.png">
     <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1" />
 
     <title>Sistem Informasi Manajemen Risiko Proyek</title>
@@ -17,24 +17,24 @@ if (!empty($this->session->userdata('id_pengguna'))) {
 
 
     <!-- Bootstrap core CSS     -->
-    <link href="<?php echo base_url(); ?>/assets/css/bootstrap.min.css" rel="stylesheet" />
+    <link href="<?php echo base_url(); ?>assets/css/bootstrap.min.css" rel="stylesheet" />
 
     <!-- Animation library for notifications   -->
-    <link href="<?php echo base_url(); ?>/assets/css/animate.min.css" rel="stylesheet"/>
+    <link href="<?php echo base_url(); ?>assets/css/animate.min.css" rel="stylesheet"/>
 
     <!--  Paper Dashboard core CSS    -->
-    <link href="<?php echo base_url(); ?>/assets/css/paper-dashboard.css" rel="stylesheet"/>
+    <link href="<?php echo base_url(); ?>assets/css/paper-dashboard.css" rel="stylesheet"/>
 
     <!--  CSS for Demo Purpose, don't include it in your project     -->
-    <link href="<?php echo base_url(); ?>/assets/css/demo.css" rel="stylesheet" />
+    <link href="<?php echo base_url(); ?>assets/css/demo.css" rel="stylesheet" />
 
     <!--  Fonts and icons     -->
     <link href="http://maxcdn.bootstrapcdn.com/font-awesome/latest/css/font-awesome.min.css" rel="stylesheet">
     <link href='https://fonts.googleapis.com/css?family=Muli:400,300' rel='stylesheet' type='text/css'>
-    <link href="<?php echo base_url(); ?>/assets/css/themify-icons.css" rel="stylesheet">
+    <link href="<?php echo base_url(); ?>assets/css/themify-icons.css" rel="stylesheet">
     
     <!--  JavaScript  -->
-    <script src="<?php echo base_url(); ?>/assets/js/jquery-3.2.1.min.js" type="text/javascript"></script>
+    <script src="<?php echo base_url(); ?>assets/js/jquery-3.2.1.min.js" type="text/javascript"></script>
 
 </head>
 <body>
@@ -51,20 +51,20 @@ if (!empty($this->session->userdata('id_pengguna'))) {
             <div class="logo">
                 <a href="#" class="simple-text">
                     <span style="line-height:30px; font-size:32px">ANANTA</span>
-                    <img src="agpp.png" style="vertical-align:bottom; width: 48px; height: 48px">
+                    <img src="<?php echo base_url(); ?>agpp.png" style="vertical-align:bottom; width: 48px; height: 48px">
                 </a>
             </div>
 
             <ul class="nav">
                 <li>
-                    <a href="profil">
+                    <a href="<?php echo base_url(); ?>profil">
                         <i class="ti-user"></i>
                         <p>Profil</p>
                     </a>
                 </li>
                 <?php if ($this->session->userdata('jabatan') == 'Admin') { ?>
                 <li>
-                    <a href="pengguna">
+                    <a href="<?php echo base_url(); ?>pengguna">
                         <i class="ti-id-badge"></i>
                         <p>Pengguna</p>
                     </a>
@@ -76,39 +76,39 @@ if (!empty($this->session->userdata('id_pengguna'))) {
                     </a>
                 </li>
                 <li>
-                    <a href="proyek">
+                    <a href="<?php echo base_url(); ?>proyek">
                         <i class="ti-agenda"></i>
                         <p>Proyek</p>
                     </a>
                 </li>
                 <?php } elseif ($this->session->userdata('jabatan') == 'Manajer Proyek') { ?>
                 <li>
-                    <a href="pekerjaan">
+                    <a href="<?php echo base_url(); ?>pekerjaan">
                         <i class="ti-list"></i>
                         <p>Pekerjaan</p>
                     </a>
                 </li>
                 <li>
-                    <a href="risiko">
+                    <a href="<?php echo base_url(); ?>risiko">
                         <i class="ti-alert"></i>
                         <p>Risiko</p>
                     </a>
                 </li>
                 <li>
-                    <a href="mitigasi">
+                    <a href="<?php echo base_url(); ?>mitigasi">
                         <i class="ti-shield"></i>
                         <p>Mitigasi</p>
                     </a>
                 </li>
                 <li>
-                    <a href="evaluasi">
+                    <a href="<?php echo base_url(); ?>evaluasi">
                         <i class="ti-write"></i>
                         <p>Evaluasi</p>
                     </a>
                 </li>
                 <?php } elseif ($this->session->userdata('jabatan') == 'Site Manager') { ?>
                 <li>
-                    <a href="laporan">
+                    <a href="<?php echo base_url(); ?>laporan">
                         <i class="ti-pencil-alt"></i>
                         <p>Laporan</p>
                     </a>
@@ -133,7 +133,7 @@ if (!empty($this->session->userdata('id_pengguna'))) {
                 <div class="collapse navbar-collapse">
                     <ul class="nav navbar-nav navbar-right">
                         <li>
-                            <a href="keluar">
+                            <a href="<?php echo base_url(); ?>keluar">
                                 <i class="ti-power-off"></i>
                                 <p>Keluar</p>
                             </a>
@@ -188,7 +188,7 @@ if (!empty($this->session->userdata('id_pengguna'))) {
                                                     <td><?php echo $klien_item['perusahaan']; ?></td>
                                                     <td><?php echo $klien_item['alamat']; ?></td>
                                                     <td><a href="klien-lihat/<?php echo $klien_item['id_klien']?>"><i class="ti-eye"></i></a></td>
-                                                    <td><a href="klien-hapus/<?php echo $klien_item['id_klien']?>"><i class="ti-trash"></i></a></td>
+                                                    <td><a href="klien-hapus/<?php echo $klien_item['id_klien']?>"><i class="ti-trash" onclick="return confirm('Anda yakin ingin menghapus data ini?')"></i></a></td>
                                                 </tr>
                                             <?php $i++; }
                                         ?>
@@ -214,24 +214,44 @@ if (!empty($this->session->userdata('id_pengguna'))) {
 
 </body>
 
+    <?php if ($success = $this->session->flashdata('success')) { ?>
+    <!--  Success Message  -->
+    <script type="text/javascript">
+        $(document).ready(function() {
+            function show_message() {
+                $.notify({
+                    icon: 'ti-info-alt',
+                    message: '<?php echo $success; ?>'
+                },{
+                    type: 'success',
+                    timer: 200
+                });
+               return false;
+            }
+
+            show_message();
+        });
+    </script>
+    <?php } ?>
+
     <!--   Core JS Files   -->
-    <script src="<?php echo base_url(); ?>/assets/js/jquery-1.10.2.js" type="text/javascript"></script>
-    <script src="<?php echo base_url(); ?>/assets/js/bootstrap.min.js" type="text/javascript"></script>
+    <script src="<?php echo base_url(); ?>assets/js/jquery-1.10.2.js" type="text/javascript"></script>
+    <script src="<?php echo base_url(); ?>assets/js/bootstrap.min.js" type="text/javascript"></script>
 
     <!--  Checkbox, Radio & Switch Plugins -->
-    <script src="<?php echo base_url(); ?>/assets/js/bootstrap-checkbox-radio.js"></script>
+    <script src="<?php echo base_url(); ?>assets/js/bootstrap-checkbox-radio.js"></script>
 
     <!--  Charts Plugin -->
-    <script src="<?php echo base_url(); ?>/assets/js/chartist.min.js"></script>
+    <script src="<?php echo base_url(); ?>assets/js/chartist.min.js"></script>
 
     <!--  Notifications Plugin    -->
-    <script src="<?php echo base_url(); ?>/assets/js/bootstrap-notify.js"></script>
+    <script src="<?php echo base_url(); ?>assets/js/bootstrap-notify.js"></script>
 
     <!-- Paper Dashboard Core javascript and methods for Demo purpose -->
-    <script src="<?php echo base_url(); ?>/assets/js/paper-dashboard.js"></script>
+    <script src="<?php echo base_url(); ?>assets/js/paper-dashboard.js"></script>
 
     <!-- Paper Dashboard DEMO methods, don't include it in your project! -->
-    <script src="<?php echo base_url(); ?>/assets/js/demo.js"></script>
+    <script src="<?php echo base_url(); ?>assets/js/demo.js"></script>
 
 
 </html>
